Fixed RT.to_JSON() and added effective_update
Function RT.to_JSON() now works again by using a slightly modified version of the Set class that you can see at the beginning of main.py.

effective_update is an addition for the real time part that holds the timestamp of the last time each trip arrived at a stop. By this way we can notify the user if a bus that was expected to take 2 minutes from stop A to stop B is taking more than 5 minute and therefore tell him that there's probably something wrong. This is done by storing the timestamp for every trip and storing a ratio for every expected stop_time. This ratio is calculated by seconds elapsed since last arrival at a stop divided by expected time between last stop and next stop. In the tg.py module if this ratio is greater equal than 3 it will notify the user that that expected time is not precise.